+++ internal/firewall/handler/errors.go
@@ -1,4 +1,4 @@
-package firewallHandler
+package handler
 
 import "errors"
